手工用例优化&&修复部分已知问题 1、手工任务适配任务组概念，优化手工用例操作流程 2、模版优化区分手工和自动化 3、优化服务重启后swagger.yaml被删除情况 4、用例执行记录兼容手工用例 5、修复用户申请加入用户组，在未通过的情况下可以查看相应的组数据

+++ radiaTest-server/server/apps/manualjob/routes.py
@@ -18,21 +18,28 @@ from flask import jsonify
 from flask_pydantic import validate
 from flask_restful import Resource
 
-from server import db, swagger_adapt
+from server import swagger_adapt
 from server.apps.manualjob.handler import (
     ManualJobHandler,
     ManualJobSubmitHandler,
     ManualJobLogHandler,
     ManualJobDeleteHandler,
-    ManualJobLogQueryHandler)
-from server.model.manualjob import ManualJob
+    ManualJobLogQueryHandler,
+    ManualJobGroupCopyHandler
+)
+from server.model.manualjob import ManualJob, ManualJobGroup
 from server.model.testcase import Case
 from server.model.milestone import Milestone
 from server.schema.manualjob import (
     ManualJobCreate,
     ManualJobQuery,
     ManualJobLogModify,
-    ManualJobLogDelete
+    ManualJobLogDelete,
+    ManualJobGroupQuery,
+    ManualJobGroupStatus,
+    ManualJobGroupReport,
+    ManualJobModify,
+    ManualJobGroupCopySchema
 )
 from server.utils.auth_util import auth
 from server.utils.response_util import response_collect, RET, workspace_error_collect
@@ -43,6 +50,17 @@ def get_manual_job_tag():
         "name": "手工任务",
         "description": "手工任务接口",
     }
+
+
+# 若任务组已完成，则上锁不允许用户进行操作
+def manual_job_operate_lock(manual_job_id):
+    manual_job = ManualJob.query.get(manual_job_id)
+    if manual_job.job_group_id:
+        # 按id查找ManualJobGroup
+        manual_job_group = ManualJobGroup.query.get(manual_job.job_group_id)
+        if manual_job_group.status == 1:
+            return False
+    return True
 
 
 class ManualJobEvent(Resource):
@@ -92,6 +110,13 @@ class ManualJobEvent(Resource):
                 error_code=RET.NO_DATA_ERR,
                 error_msg=f"the milestone with id {body.milestone_id} does not exist"
             )
+        _manual_job_group = ManualJobGroup.query.filter_by(name=body.name).all()
+        if _manual_job_group:
+            return jsonify(
+                error_code=RET.NO_DATA_ERR,
+                error_msg=f"任务名重复，请修改后重试！"
+            )
+
         return ManualJobHandler.create(body)
 
     @auth.login_required()
@@ -170,7 +195,6 @@ class ManualJobSubmitEvent(Resource):
         "summary": "设置手工任务状态已结束(每个步骤都有日志, 且日志的passed字段都是True,"
                    " 则把此手工测试任务的result字段更改为1(与预期一致))",
         "externalDocs": {"description": "", "url": ""},  # 当前接口扩展文档定义
-        "request_schema_model": ManualJobQuery
     })
     def post(self, manual_job_id: int):
         """
@@ -183,6 +207,11 @@ class ManualJobSubmitEvent(Resource):
                 "error_msg": "Request processed successfully."
             }
         """
+        if manual_job_operate_lock(manual_job_id) is False:
+            return jsonify(
+                error_code=RET.UNAUTHORIZED_ACCESS,
+                error_msg=f"该任务已被锁定，请将状态改为执行中后再进行该操作！"
+            )
         manual_job = ManualJob.query.get(manual_job_id)
         if manual_job is None:
             return jsonify(
@@ -191,6 +220,33 @@ class ManualJobSubmitEvent(Resource):
             )
         else:
             return ManualJobSubmitHandler.post(manual_job)
+
+    @auth.login_required()
+    @response_collect
+    @validate()
+    @swagger_adapt.api_schema_model_map({
+        "__module__": get_manual_job_tag.__module__,   # 获取当前接口所在模块
+        "resource_name": "ManualJobSubmitEvent",  # 当前接口视图函数名
+        "func_name": "put",   # 当前接口所对应的函数名
+        "tag": get_manual_job_tag(),  # 当前接口所对应的标签
+        "summary": "设置手工任务结果",
+        "externalDocs": {"description": "", "url": ""},  # 当前接口扩展文档定义
+        "request_schema_model": ManualJobModify
+    })
+    def put(self, manual_job_id: int, body: ManualJobModify):
+        if manual_job_operate_lock(manual_job_id) is False:
+            return jsonify(
+                error_code=RET.UNAUTHORIZED_ACCESS,
+                error_msg=f"该任务已被锁定，请将状态改为执行中后再进行该操作！"
+            )
+        manual_job = ManualJob.query.get(manual_job_id)
+        if manual_job is None:
+            return jsonify(
+                error_code=RET.NO_DATA_ERR,
+                error_msg=f"the manual_job with id {manual_job_id} does not exist"
+            )
+        else:
+            return ManualJobSubmitHandler.update(manual_job, body)
 
 
 class ManualJobLogEvent(Resource):
@@ -226,6 +282,11 @@ class ManualJobLogEvent(Resource):
                 "error_msg": "Request processed successfully."
             }
         """
+        if manual_job_operate_lock(manual_job_id) is False:
+            return jsonify(
+                error_code=RET.UNAUTHORIZED_ACCESS,
+                error_msg=f"该任务已被锁定，请将状态改为执行中后再进行该操作！"
+            )
         return ManualJobLogHandler.update(manual_job_id, body)
     
     @auth.login_required()
@@ -254,6 +315,11 @@ class ManualJobLogEvent(Resource):
                 "error_msg": "Request processed successfully."
             }
         """
+        if manual_job_operate_lock(manual_job_id) is False:
+            return jsonify(
+                error_code=RET.UNAUTHORIZED_ACCESS,
+                error_msg=f"该任务已被锁定，请将状态改为执行中后再进行该操作！"
+            )
         return ManualJobLogHandler.delete(manual_job_id, body)
 
 
@@ -283,6 +349,11 @@ class ManualJobDeleteEvent(Resource):
                 "error_msg": "Request processed successfully."
             }
         """
+        if manual_job_operate_lock(manual_job_id) is False:
+            return jsonify(
+                error_code=RET.UNAUTHORIZED_ACCESS,
+                error_msg=f"该任务已被锁定，请将状态改为执行中后再进行该操作！"
+            )
         return ManualJobDeleteHandler.delete(manual_job_id)
 
 
@@ -325,3 +396,150 @@ class ManualJobLogQueryEvent(Resource):
                 error_msg=f"manual_job with id {manual_job_id} does not exist"
             )
         return ManualJobLogQueryHandler.query(manual_job, step_num)
+
+
+class ManualJobGroupEvent(Resource):
+    @auth.login_required()
+    @response_collect
+    @workspace_error_collect
+    @validate()
+    @swagger_adapt.api_schema_model_map({
+        "__module__": get_manual_job_tag.__module__,   # 获取当前接口所在模块
+        "resource_name": "ManualJobGroupEvent",  # 当前接口视图函数名
+        "func_name": "get",   # 当前接口所对应的函数名
+        "tag": get_manual_job_tag(),  # 当前接口所对应的标签
+        "summary": "分页查询手工任务组",  # 当前接口概述
+        "externalDocs": {"description": "", "url": ""},  # 当前接口扩展文档定义
+        "query_schema_model": ManualJobGroupQuery
+    })
+    def get(self, workspace: str, query: ManualJobGroupQuery):
+        return ManualJobHandler.query_manual_job_group(query, workspace)
+
+
+class ManualJobGroupDetail(Resource):
+    @auth.login_required()
+    @response_collect
+    @workspace_error_collect
+    @swagger_adapt.api_schema_model_map({
+        "__module__": get_manual_job_tag.__module__,   # 获取当前接口所在模块
+        "resource_name": "ManualJobGroupDetail",  # 当前接口视图函数名
+        "func_name": "get",   # 当前接口所对应的函数名
+        "tag": get_manual_job_tag(),  # 当前接口所对应的标签
+        "summary": "手工任务组详情",  # 当前接口概述
+        "externalDocs": {"description": "", "url": ""},  # 当前接口扩展文档定义
+    })
+    def get(self, manual_job_group_id):
+        # 按id查找ManualJobGroup
+        manual_job_group = ManualJobGroup.query.get(manual_job_group_id)
+        if manual_job_group is None:
+            return jsonify(
+                error_code=RET.NO_DATA_ERR,
+                error_msg=f"manual_job_group with id {manual_job_group_id} does not exist"
+            )
+        return jsonify(
+            data=manual_job_group.to_statist(),
+            error_code=RET.OK,
+            error_msg="Request processed successfully."
+            )
+
+    @auth.login_required()
+    @response_collect
+    @workspace_error_collect
+    @validate()
+    @swagger_adapt.api_schema_model_map({
+        "__module__": get_manual_job_tag.__module__,   # 获取当前接口所在模块
+        "resource_name": "ManualJobGroupDetail",  # 当前接口视图函数名
+        "func_name": "put",   # 当前接口所对应的函数名
+        "tag": get_manual_job_tag(),  # 当前接口所对应的标签
+        "summary": "手工任务组状态变更",  # 当前接口概述
+        "externalDocs": {"description": "", "url": ""},  # 当前接口扩展文档定义
+        "request_schema_model": ManualJobGroupStatus
+    })
+    def put(self, manual_job_group_id, body: ManualJobGroupStatus):
+        # 按id查找ManualJobGroup
+        manual_job_group = ManualJobGroup.query.get(manual_job_group_id)
+        if manual_job_group is None:
+            return jsonify(
+                error_code=RET.NO_DATA_ERR,
+                error_msg=f"manual_job_group with id {manual_job_group_id} does not exist"
+            )
+        if manual_job_group.status == body.status:
+            return jsonify(
+                        error_code=RET.DATA_EXIST_ERR,
+                        error_msg="任务状态已经变更，无需重复操作"
+                    )
+
+        manual_job_group.status = body.status
+        manual_job_group.add_update()
+        return jsonify(
+            error_code=RET.OK,
+            error_msg="Request processed successfully."
+        )
+
+    @auth.login_required()
+    @response_collect
+    @workspace_error_collect
+    @validate()
+    @swagger_adapt.api_schema_model_map({
+        "__module__": get_manual_job_tag.__module__,   # 获取当前接口所在模块
+        "resource_name": "ManualJobGroupDetail",  # 当前接口视图函数名
+        "func_name": "post",   # 当前接口所对应的函数名
+        "tag": get_manual_job_tag(),  # 当前接口所对应的标签
+        "summary": "手工任务组报告上传",  # 当前接口概述
+        "externalDocs": {"description": "", "url": ""},  # 当前接口扩展文档定义
+        "request_schema_model": ManualJobGroupReport
+    })
+    def post(self, manual_job_group_id, body: ManualJobGroupReport):
+        # 按id查找ManualJobGroup
+        manual_job_group = ManualJobGroup.query.get(manual_job_group_id)
+        if manual_job_group is None:
+            return jsonify(
+                error_code=RET.NO_DATA_ERR,
+                error_msg=f"manual_job_group with id {manual_job_group_id} does not exist"
+            )
+
+        manual_job_group.report = body.report
+        manual_job_group.add_update()
+        return jsonify(
+            error_code=RET.OK,
+            error_msg="Request processed successfully."
+        )
+
+    @auth.login_required()
+    @response_collect
+    @workspace_error_collect
+    @swagger_adapt.api_schema_model_map({
+        "__module__": get_manual_job_tag.__module__,   # 获取当前接口所在模块
+        "resource_name": "ManualJobGroupDetail",  # 当前接口视图函数名
+        "func_name": "delete",   # 当前接口所对应的函数名
+        "tag": get_manual_job_tag(),  # 当前接口所对应的标签
+        "summary": "手工任务组删除",  # 当前接口概述
+        "externalDocs": {"description": "", "url": ""},  # 当前接口扩展文档定义
+    })
+    def delete(self, manual_job_group_id):
+        return ManualJobDeleteHandler.delete_manual_job_group(manual_job_group_id)
+
+
+class ManualJobGroupCopy(Resource):
+    @auth.login_required()
+    @response_collect
+    @workspace_error_collect
+    @validate()
+    @swagger_adapt.api_schema_model_map({
+        "__module__": get_manual_job_tag.__module__,  # 获取当前接口所在模块
+        "resource_name": "ManualJobGroupCopy",  # 当前接口视图函数名
+        "func_name": "post",  # 当前接口所对应的函数名
+        "tag": get_manual_job_tag(),  # 当前接口所对应的标签
+        "summary": "复制手工任务组",  # 当前接口概述
+        "externalDocs": {"description": "", "url": ""},  # 当前接口扩展文档定义
+        "request_schema_model": ManualJobGroupCopySchema
+    })
+    def post(self, body: ManualJobGroupCopySchema):
+        # 按id查找ManualJobGroup
+        manual_job_group = ManualJobGroup.query.get(body.id)
+        if manual_job_group is None:
+            return jsonify(
+                error_code=RET.NO_DATA_ERR,
+                error_msg=f"manual_job_group with id {body.id} does not exist"
+            )
+        return ManualJobGroupCopyHandler.copy(manual_job_group, body.__dict__)
